// src/app/modules/admin/apps/orders/orders.types.ts
import { Channel } from "app/layout/common/channel/all-channels.types";
import { Application } from "../applications/applications.types";
import { Campaign } from "../campaigns/campaigns.types";

export interface Order {
action_taken: any;
stage: any;
comment: any;
job_description: any;
video_script_desc: any;
    id: string;
    order_id:string
    bidAmount: string | null;
    channelName: string | null;
    deadlineDate: string;
    brandTnc: string;
    tncDisagreeReason: string;
    secondPreference: string;
    secondPreferenceAlt: string;
    firstHalf: string;
    firstHalfAlt: string;
    videoReview: string;
    minimumVideo: string;
    minimumVideoCount: number;
    changesFlexibility: string;
    goLiveTime: string;
    goLiveTimeOther: number;
    deliveryTime: string;
    delivery_days: string;
    deliveryTimeOther: number;
    shareOnSocial: string;
    social_media_share: string;
    facebook: boolean;
    instagram: boolean;
    twitter: boolean;
    shareUnlistedPreview: boolean;
    nonViolationPolicy: boolean;
    userComment: string | null;
    status: string | null;
    disputeReason: string;
    channel: Channel;
    channel_info: any;
    instagram_data: any;
    orderStatus: string[];
    currency: string;
    camp_info: Campaign;
    app_info: Application;
    pay_amount: number;
    camp_price: number;
    total_pay: number;
    minor_changes: number;
    payment_term: number;
    campaign: Campaign;
    order_process: any;
    instruction: string;
    features: string;
    suggestions: string;
    template_script: string;
    promotion_part: string;
    messageCount: number;
}